requirements.txt
biopython==1.79
click==8.1.3
Flask==2.1.2
Flask-WTF==1.0.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
typing_extensions==4.2.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
